requirements.txt
streamlit
sklearn
matplotlib
opencv-python
Pillow
numpy